+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 All notable changes to this project will be documented in this file.
 
 #### 0.0.x Releases
-- `0.0.15` Releases - [0.0.15](#0014)
+- `0.0.16` Releases - [0.0.16](#0016)
+- `0.0.15` Releases - [0.0.15](#0015)
 - `0.0.13` Releases - [0.0.13](#0013)
 - `0.0.12` Releases - [0.0.12](#0012)
 - `0.0.9` Releases - [0.0.9](#009)
@@ -13,6 +14,13 @@ All notable changes to this project will be documented in this file.
 - `0.0.3` Releases - [0.0.3](#003)
 - `0.0.2` Releases - [0.0.2](#002)
 - `0.0.1` Releases - [0.0.1](#001)
+
+## [0.0.16]
+Released on 2018-12-29
+
+#### Added
+- `UIWindow` extension add topController variable.
+- `UIViewController` extension add modalPresent method.
 
 ## [0.0.15]
 Released on 2018-12-25